Move StepLayout up one directory level
